Split out test-utils generic strategies into their own crate
